requirements.txt
SQLAlchemy==1.4.46
python-dotenv==1.0.0
mysql-connector-python==4.21.12